backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=3.0.0
redis>=4.3.4
google-generativeai>=0.3.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings
better-profanity>=0.7.0
